+++ doc/examples/test_examples.py
@@ -52,7 +52,9 @@ def find_test_cases():
         basename = os.path.basename(script)
         name, _ = os.path.splitext(basename)
         for format in "txt cobertura csv json html".split():
-            if format == "html" and is_compiler(os.getenv("CC"), "gcc-5", "gcc-6"):
+            if format == "html" and is_compiler(
+                os.getenv("CC"), "gcc-5", "gcc-6", "gcc-14"
+            ):
                 continue
             baseline = "{data_dirname}/{name}.{ext}".format(
                 data_dirname=data_dirname,